Added Product Update Feature

+++ frutility-backend/Controllers/ProductsController.cs
@@ -225,12 +225,10 @@ namespace frutility_backend.Controllers
         //Post: api/products
         [Authorize]
         [HttpPut("{id}")]
-        public async Task<ActionResult<Products>> PostProduct([FromForm]ProductUpdateVM model)
+        public async Task<ActionResult<Products>> UpdateProduct(long id,[FromForm]ProductUpdateVM model)
         {
-            if (!ModelState.IsValid)
-            {
+            if (id != model.Id)
                 return BadRequest("Invalid Data");
-            }
             var product = await _context.Products.FirstOrDefaultAsync(p => p.Id == model.Id);
             string UniqueFileNameimage1 = null;
             if (model.ImageNo1)
@@ -240,8 +238,11 @@ namespace frutility_backend.Controllers
                     string uploadfolder = Path.Combine(_hostingEnvironment.ContentRootPath, "Assests/images");
                     UniqueFileNameimage1 = Guid.NewGuid().ToString() + "_" + model.Image1.FileName;
                     string filepath = Path.Combine(uploadfolder, UniqueFileNameimage1);
-                    string prevFile = Path.Combine(uploadfolder, product.Image1);
-                    System.IO.File.Delete(prevFile);
+                    if (product.Image1 != null)
+                    {
+                        string prevFile = Path.Combine(uploadfolder, product.Image1);
+                        System.IO.File.Delete(prevFile);
+                    }
                     using (var stream = System.IO.File.Create(filepath))
                     {
                         await model.Image1.CopyToAsync(stream);
@@ -256,8 +257,11 @@ namespace frutility_backend.Controllers
                     string uploadfolder = Path.Combine(_hostingEnvironment.ContentRootPath, "Assests/images");
                     UniqueFileNameimage2 = Guid.NewGuid().ToString() + "_" + model.Image2.FileName;
                     string filepath = Path.Combine(uploadfolder, UniqueFileNameimage2);
-                    string prevFile = Path.Combine(uploadfolder, product.Image2);
-                    System.IO.File.Delete(prevFile);
+                    if (product.Image2 != null)
+                    {
+                        string prevFile = Path.Combine(uploadfolder, product.Image2);
+                        System.IO.File.Delete(prevFile);
+                    }
                     using (var stream = System.IO.File.Create(filepath))
                     {
                         await model.Image2.CopyToAsync(stream);
@@ -273,51 +277,37 @@ namespace frutility_backend.Controllers
                     string uploadfolder = Path.Combine(_hostingEnvironment.ContentRootPath, "Assests/images");
                     UniqueFileNameimage3 = Guid.NewGuid().ToString() + "_" + model.Image3.FileName;
                     string filepath = Path.Combine(uploadfolder, UniqueFileNameimage3);
-                    string prevFile = Path.Combine(uploadfolder, product.Image3);
-                    System.IO.File.Delete(prevFile);
+                    if(product.Image3 != null)
+                    {
+                        string prevFile = Path.Combine(uploadfolder, product.Image3);
+                        System.IO.File.Delete(prevFile);
+                    }
                     using (var stream = System.IO.File.Create(filepath))
                     {
                         await model.Image3.CopyToAsync(stream);
                     }
                 }
             }
-
-            Products products = new Products
-            {
-                Id = model.Id,
-                Name = model.Name,
-                Description = model.Description,
-                Vendor = model.Vendor,
-                Price = model.Price,
-                PriceBeforeDiscount = model.PriceBeforeDiscount,
-                Image1 = (UniqueFileNameimage1 != null) ? UniqueFileNameimage1 : product.Image1,
-                Image2 = (UniqueFileNameimage2 != null) ? UniqueFileNameimage2 : product.Image2,
-                Image3 = (UniqueFileNameimage3 != null) ? UniqueFileNameimage3 : product.Image3,
-                ShippingCharges = model.ShippingCharges,
-                Availability = model.Availability,
-                Stock = model.Stock,
-                PostingDate = DateTime.Now,
-                UpdationDate = DateTime.Now,
-                PackageWeight = model.PackageWeight,
-                SubCategoryID = model.SubCategoryID
-            };
-            _context.Products.Add(products);
+            product.Id = model.Id;
+            product.Name = model.Name;
+            product.Description = model.Description;
+            product.Vendor = model.Vendor;
+            product.Price = model.Price;
+            product.PriceBeforeDiscount = model.PriceBeforeDiscount;
+            product.Image1 = (UniqueFileNameimage1 != null) ? UniqueFileNameimage1 : product.Image1;
+            product.Image2 = (UniqueFileNameimage2 != null) ? UniqueFileNameimage2 : product.Image2;
+            product.Image3 = (UniqueFileNameimage3 != null) ? UniqueFileNameimage3 : product.Image3;
+            product.ShippingCharges = model.ShippingCharges;
+            product.Availability = model.Availability;
+            product.Stock = model.Stock;
+            product.UpdationDate = DateTime.Now;
+            product.PackageWeight = model.PackageWeight;
+            product.SubCategoryID = model.SubCategoryID;
+            _context.Entry(product).State = EntityState.Modified;
             await _context.SaveChangesAsync();
-            return Ok(products);
+            return Ok(product);
         }
 
-        //Put: api/products/{id}
-        [HttpPut("{id}")]
-        public async Task<ActionResult<Products>> UpdateProducts(int id, Products productsrec)
-        {
-            if(id != productsrec.Id)
-            {
-                return BadRequest();
-            }
-            _context.Entry(productsrec).State = EntityState.Modified;
-            await _context.SaveChangesAsync();
-            return NoContent();
-        }
 
         //Delete: api/products/{id}
         [HttpDelete("{id}")]
